Ingame menus (#14)
* feat(ingame-menus) : Added character status menu and fixed InputBox lisibility

* feat(ingame-menus) : Added character status menu and fixed InputBox lisibility

* feat(ingame-menus) : graphical improvements

* feat(ingame-menus) : Added character status menu and fixed InputBox lisibility

* feat(ingame-menus) : Added character status menu and fixed InputBox lisibility

* feat(ingame-menus) : graphical improvements

* exceptionnal(ingame-menus) : pushing stuff that doesn't work in for personal purposes

* feat(ingame-menus)à : fully working inventory, character status sheet and items

* feat(ingame-menus): aligned nickname and exported colors

* pull-request(ingame-menus) : adjustments and corrections for the pr reviews

* feat(ingame-menus): nickname now blitted only once

+++ src/menu/MenuCharacter.py
@@ -2,7 +2,7 @@
 import os
 import signal
 from src.menu._menu import Menu
-from src.interface import Button
+from src.interface import Button, Text, Input
 from src.config.assets import fonts_folder, menus_folder
 from os import path
 from src.config.colors import WHITE
@@ -19,11 +19,17 @@ class MenuCharacter(Menu):
             path.join(menus_folder, "background.png")
         ).convert_alpha()
 
+        self.name_text = Text(self.game.display, self.game.resolution[0] // 2, self.game.resolution[1]
+                              // 2 - 15, "Enter your character name :", path.join(fonts_folder, "CascadiaCode.ttf"), WHITE, 15, True)
+
+        self.name_input = Input(self.game, self.game.resolution[0] // 2, self.game.resolution[1]
+                                // 2 + 25, 350, 40, path.join(fonts_folder, "CascadiaCode.ttf"), 20, 15)
+
         self.play_button = Button(
             self.game,
             self.game.resolution[0] // 2,
-            self.game.resolution[1] // 2,
-            "Play",
+            self.game.resolution[1] // 2 + 150,
+            "Create game",
             path.join(fonts_folder, "enchanted_land.otf"),
             WHITE,
             40,
@@ -31,8 +37,8 @@ class MenuCharacter(Menu):
         self.join_button = Button(
             self.game,
             self.game.resolution[0] // 2,
-            self.game.resolution[1] // 2 + 150,
-            "Join",
+            self.game.resolution[1] // 2 + 300,
+            "Join game",
             path.join(fonts_folder, "enchanted_land.otf"),
             WHITE,
             40,
@@ -53,18 +59,32 @@ class MenuCharacter(Menu):
             if isinstance(getattr(self, button), Button)
         ]
 
+        self.texts = [
+            getattr(self, text)
+            for text in dir(self)
+            if isinstance(getattr(self, text), Text)
+        ]
+
+        self.inputs = [
+            getattr(self, input_box)
+            for input_box in dir(self)
+            if isinstance(getattr(self, input_box), Input)
+        ]
+
     def check_events(self, event):
         """This method deals with the user inputs
             on the different buttons of the menu
         Args:
             event (Event): a pygame event
         """
-        if self.play_button.is_clicked(event):
+        self.name_input.handle_events(event)
+        if self.play_button.is_clicked(event) and self.name_input.get_text():
             self.game.menu_running = False
             self.game.playing = True
             # unpause the server
             os.kill(self.game.network._server.pid, signal.SIGUSR2)
             self.displaying = False
+            self.game.player.set_nickname(self.name_input.get_text())
         if self.join_button.is_clicked(event):
             self.game.current_menu = self.game.join_menu
             self.displaying = False
@@ -83,6 +103,12 @@ class MenuCharacter(Menu):
             for button in self.buttons:
                 button.display_button()
                 button.color_on_mouse(WHITE)
+
+            for input_box in self.inputs:
+                input_box.display_box()
+
+            for text in self.texts:
+                text.display_text()
 
             self.display_to_game()
             self.game.clock.tick(30)
